+++ R/gess_fisher.R
@@ -45,6 +45,11 @@
 #' \code{lower} argument need to be set as \code{NULL} if the \code{refdb} in \code{qSig} 
 #' is path to the HDF5 file that were converted from the gmt file.
 #' 
+#' @param padj numeric(1), cutoff of adjusted p-value or false discovery rate (FDR)
+#' of defining DEGs that is less than or equal to 'padj'. The 'padj' argument is
+#' valid only if the reference HDF5 file contains the p-value matrix stored in 
+#' the dataset named as 'padj'. 
+#' 
 #' @param chunk_size number of database entries to process per iteration to 
 #' limit memory usage of search.
 #' @param ref_trts character vector. If users want to search against a subset 
@@ -80,8 +85,8 @@
 #' # result(fisher)
 #' @export
 #' 
-gess_fisher <- function(qSig, higher=NULL, lower=NULL, chunk_size=5000, 
-                        ref_trts=NULL, workers=1){
+gess_fisher <- function(qSig, higher=NULL, lower=NULL, padj=NULL,
+                        chunk_size=5000, ref_trts=NULL, workers=1){
   if(!is(qSig, "qSig")) stop("The 'qSig' should be an object of 'qSig' class")
   #stopifnot(validObject(qSig))
   if(gm(qSig) != "Fisher"){
@@ -108,11 +113,27 @@ gess_fisher <- function(qSig, higher=NULL, lower=NULL, chunk_size=5000,
   
   full_dim <- dim(full_mat)
   full_grid <- colGrid(full_mat, ncol=min(chunk_size, ncol(full_mat)))
+  
+  if(! is.null(padj)){
+      if(! 'padj' %in% h5ls(db_path)$name){
+          stop("The 'refdb' need to be an hdf5 file that contains 'padj' dataset!")
+      }
+      full_pmat <- HDF5Array(db_path, name="padj")
+      rownames(full_pmat) <- as.character(HDF5Array(db_path, name="rownames"))
+      colnames(full_pmat) <- as.character(HDF5Array(db_path, name="colnames"))
+  }
+  
   ### The blocks in 'full_grid' are made of full columns 
   nblock <- length(full_grid) 
   resultDF <- bplapply(seq_len(nblock), function(b){
     ref_block <- read_block(full_mat, full_grid[[b]])
-    cmap <- induceSMat(ref_block, higher=higher, lower=lower)
+    if(! is.null(padj)){
+        pmat <- read_block(full_pmat, full_grid[[b]])
+    } else {
+        pmat = NULL
+    }
+    cmap <- induceSMat(ref_block, higher=higher, lower=lower,
+                       padj=padj, pmat=pmat)
     universe <- rownames(cmap)
     c <- fs(query=query, sets=cmap, universe = universe)
     return(data.frame(c))}, BPPARAM = MulticoreParam(workers = workers))